+++ client/driver/docker_test.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"io/ioutil"
 	"math/rand"
-	"os"
-	"os/exec"
 	"path/filepath"
 	"reflect"
 	"runtime/debug"
@@ -15,13 +13,10 @@ import (
 	"time"
 
 	docker "github.com/fsouza/go-dockerclient"
-	"github.com/hashicorp/go-plugin"
 	"github.com/hashicorp/nomad/client/allocdir"
 	"github.com/hashicorp/nomad/client/config"
 	"github.com/hashicorp/nomad/client/driver/env"
-	cstructs "github.com/hashicorp/nomad/client/driver/structs"
 	"github.com/hashicorp/nomad/client/testutil"
-	"github.com/hashicorp/nomad/helper/discover"
 	"github.com/hashicorp/nomad/nomad/structs"
 	tu "github.com/hashicorp/nomad/testutil"
 )
@@ -54,7 +49,10 @@ func dockerTask() (*structs.Task, int, int) {
 	return &structs.Task{
 		Name: "redis-demo",
 		Config: map[string]interface{}{
-			"image": "redis",
+			"image":   "busybox",
+			"load":    []string{"busybox.tar"},
+			"command": "/bin/nc",
+			"args":    []string{"-l", "127.0.0.1", "0"},
 		},
 		LogConfig: &structs.LogConfig{
 			MaxFiles:      10,
@@ -97,6 +95,7 @@ func dockerSetup(t *testing.T, task *structs.Task) (*docker.Client, DriverHandle
 
 	driverCtx, execCtx := testDriverContexts(task)
 	driver := NewDockerDriver(driverCtx)
+	copyImage(execCtx, task, "busybox.tar", t)
 
 	handle, err := driver.Start(execCtx, task)
 	if err != nil {
@@ -116,49 +115,8 @@ func dockerSetup(t *testing.T, task *structs.Task) (*docker.Client, DriverHandle
 	return client, handle, cleanup
 }
 
-func TestDockerDriver_Handle(t *testing.T) {
-	t.Parallel()
-
-	bin, err := discover.NomadExecutable()
-	if err != nil {
-		t.Fatalf("got an err: %v", err)
-	}
-
-	f, _ := ioutil.TempFile(os.TempDir(), "")
-	defer f.Close()
-	defer os.Remove(f.Name())
-	pluginConfig := &plugin.ClientConfig{
-		Cmd: exec.Command(bin, "syslog", f.Name()),
-	}
-	exec, pluginClient, err := createExecutor(pluginConfig, os.Stdout, &config.Config{})
-	if err != nil {
-		t.Fatalf("got an err: %v", err)
-	}
-	defer pluginClient.Kill()
-
-	h := &DockerHandle{
-		version:        "version",
-		imageID:        "imageid",
-		executor:       exec,
-		pluginClient:   pluginClient,
-		containerID:    "containerid",
-		killTimeout:    5 * time.Nanosecond,
-		maxKillTimeout: 15 * time.Nanosecond,
-		doneCh:         make(chan bool),
-		waitCh:         make(chan *cstructs.WaitResult, 1),
-	}
-
-	actual := h.ID()
-	expected := fmt.Sprintf("DOCKER:{\"Version\":\"version\",\"ImageID\":\"imageid\",\"ContainerID\":\"containerid\",\"KillTimeout\":5,\"MaxKillTimeout\":15,\"PluginConfig\":{\"Pid\":%d,\"AddrNet\":\"unix\",\"AddrName\":\"%s\"}}",
-		pluginClient.ReattachConfig().Pid, pluginClient.ReattachConfig().Addr.String())
-	if actual != expected {
-		t.Errorf("Expected `%s`, found `%s`", expected, actual)
-	}
-}
-
 // This test should always pass, even if docker daemon is not available
 func TestDockerDriver_Fingerprint(t *testing.T) {
-	t.Parallel()
 	driverCtx, _ := testDriverContexts(&structs.Task{Name: "foo"})
 	d := NewDockerDriver(driverCtx)
 	node := &structs.Node{
@@ -178,15 +136,17 @@ func TestDockerDriver_Fingerprint(t *testing.T) {
 }
 
 func TestDockerDriver_StartOpen_Wait(t *testing.T) {
-	t.Parallel()
 	if !testutil.DockerIsConnected(t) {
 		t.SkipNow()
 	}
 
 	task := &structs.Task{
-		Name: "redis-demo",
+		Name: "nc-demo",
 		Config: map[string]interface{}{
-			"image": "redis",
+			"load":    []string{"busybox.tar"},
+			"image":   "busybox",
+			"command": "/bin/nc",
+			"args":    []string{"-l", "127.0.0.1", "0"},
 		},
 		LogConfig: &structs.LogConfig{
 			MaxFiles:      10,
@@ -198,6 +158,7 @@ func TestDockerDriver_StartOpen_Wait(t *testing.T) {
 	driverCtx, execCtx := testDriverContexts(task)
 	defer execCtx.AllocDir.Destroy()
 	d := NewDockerDriver(driverCtx)
+	copyImage(execCtx, task, "busybox.tar", t)
 
 	handle, err := d.Start(execCtx, task)
 	if err != nil {
@@ -219,13 +180,13 @@ func TestDockerDriver_StartOpen_Wait(t *testing.T) {
 }
 
 func TestDockerDriver_Start_Wait(t *testing.T) {
-	t.Parallel()
 	task := &structs.Task{
-		Name: "redis-demo",
+		Name: "nc-demo",
 		Config: map[string]interface{}{
-			"image":   "redis",
-			"command": "/usr/local/bin/redis-server",
-			"args":    []string{"-v"},
+			"load":    []string{"busybox.tar"},
+			"image":   "busybox",
+			"command": "/bin/echo",
+			"args":    []string{"hello"},
 		},
 		Resources: &structs.Resources{
 			MemoryMB: 256,
@@ -257,7 +218,6 @@ func TestDockerDriver_Start_Wait(t *testing.T) {
 }
 
 func TestDockerDriver_Start_LoadImage(t *testing.T) {
-	t.Parallel()
 	if !testutil.DockerIsConnected(t) {
 		t.SkipNow()
 	}
@@ -323,7 +283,6 @@ func TestDockerDriver_Start_LoadImage(t *testing.T) {
 }
 
 func TestDockerDriver_Start_Wait_AllocDir(t *testing.T) {
-	t.Parallel()
 	// This test requires that the alloc dir be mounted into docker as a volume.
 	// Because this cannot happen when docker is run remotely, e.g. when running
 	// docker in a VM, we skip this when we detect Docker is being run remotely.
@@ -334,10 +293,11 @@ func TestDockerDriver_Start_Wait_AllocDir(t *testing.T) {
 	exp := []byte{'w', 'i', 'n'}
 	file := "output.txt"
 	task := &structs.Task{
-		Name: "redis-demo",
+		Name: "nc-demo",
 		Config: map[string]interface{}{
-			"image":   "redis",
-			"command": "/bin/bash",
+			"image":   "busybox",
+			"load":    []string{"busybox.tar"},
+			"command": "/bin/sh",
 			"args": []string{
 				"-c",
 				fmt.Sprintf(`sleep 1; echo -n %s > $%s/%s`,
@@ -357,6 +317,7 @@ func TestDockerDriver_Start_Wait_AllocDir(t *testing.T) {
 	driverCtx, execCtx := testDriverContexts(task)
 	defer execCtx.AllocDir.Destroy()
 	d := NewDockerDriver(driverCtx)
+	copyImage(execCtx, task, "busybox.tar", t)
 
 	handle, err := d.Start(execCtx, task)
 	if err != nil {
@@ -389,11 +350,11 @@ func TestDockerDriver_Start_Wait_AllocDir(t *testing.T) {
 }
 
 func TestDockerDriver_Start_Kill_Wait(t *testing.T) {
-	t.Parallel()
 	task := &structs.Task{
-		Name: "redis-demo",
+		Name: "nc-demo",
 		Config: map[string]interface{}{
-			"image":   "redis",
+			"image":   "busybox",
+			"load":    []string{"busybox.tar"},
 			"command": "/bin/sleep",
 			"args":    []string{"10"},
 		},
@@ -425,8 +386,7 @@ func TestDockerDriver_Start_Kill_Wait(t *testing.T) {
 	}
 }
 
-func TestDocker_StartN(t *testing.T) {
-	t.Parallel()
+func TestDockerDriver_StartN(t *testing.T) {
 	if !testutil.DockerIsConnected(t) {
 		t.SkipNow()
 	}
@@ -446,6 +406,7 @@ func TestDocker_StartN(t *testing.T) {
 		driverCtx, execCtx := testDriverContexts(task)
 		defer execCtx.AllocDir.Destroy()
 		d := NewDockerDriver(driverCtx)
+		copyImage(execCtx, task, "busybox.tar", t)
 
 		handles[idx], err = d.Start(execCtx, task)
 		if err != nil {
@@ -470,20 +431,22 @@ func TestDocker_StartN(t *testing.T) {
 	t.Log("Test complete!")
 }
 
-func TestDocker_StartNVersions(t *testing.T) {
-	t.Parallel()
+func TestDockerDriver_StartNVersions(t *testing.T) {
 	if !testutil.DockerIsConnected(t) {
 		t.SkipNow()
 	}
 
 	task1, _, _ := dockerTask()
-	task1.Config["image"] = "redis"
+	task1.Config["image"] = "busybox"
+	task1.Config["load"] = []string{"busybox.tar"}
 
 	task2, _, _ := dockerTask()
-	task2.Config["image"] = "redis:latest"
+	task2.Config["image"] = "busybox:musl"
+	task2.Config["load"] = []string{"busybox_musl.tar"}
 
 	task3, _, _ := dockerTask()
-	task3.Config["image"] = "redis:3.0"
+	task3.Config["image"] = "busybox:glibc"
+	task3.Config["load"] = []string{"busybox_glibc.tar"}
 
 	taskList := []*structs.Task{task1, task2, task3}
 
@@ -497,6 +460,9 @@ func TestDocker_StartNVersions(t *testing.T) {
 		driverCtx, execCtx := testDriverContexts(task)
 		defer execCtx.AllocDir.Destroy()
 		d := NewDockerDriver(driverCtx)
+		copyImage(execCtx, task, "busybox.tar", t)
+		copyImage(execCtx, task, "busybox_musl.tar", t)
+		copyImage(execCtx, task, "busybox_glibc.tar", t)
 
 		handles[idx], err = d.Start(execCtx, task)
 		if err != nil {
@@ -521,14 +487,31 @@ func TestDocker_StartNVersions(t *testing.T) {
 	t.Log("Test complete!")
 }
 
-func TestDockerHostNet(t *testing.T) {
-	t.Parallel()
+func waitForExist(t *testing.T, client *docker.Client, handle *DockerHandle) {
+	tu.WaitForResult(func() (bool, error) {
+		container, err := client.InspectContainer(handle.ContainerID())
+		if err != nil {
+			if _, ok := err.(*docker.NoSuchContainer); !ok {
+				return false, err
+			}
+		}
+
+		return container != nil, nil
+	}, func(err error) {
+		t.Fatalf("err: %v", err)
+	})
+}
+
+func TestDockerDriver_NetworkMode_Host(t *testing.T) {
 	expected := "host"
 
 	task := &structs.Task{
-		Name: "redis-demo",
+		Name: "nc-demo",
 		Config: map[string]interface{}{
-			"image":        "redis",
+			"image":        "busybox",
+			"load":         []string{"busybox.tar"},
+			"command":      "/bin/nc",
+			"args":         []string{"-l", "127.0.0.1", "0"},
 			"network_mode": expected,
 		},
 		Resources: &structs.Resources{
@@ -544,6 +527,8 @@ func TestDockerHostNet(t *testing.T) {
 	client, handle, cleanup := dockerSetup(t, task)
 	defer cleanup()
 
+	waitForExist(t, client, handle.(*DockerHandle))
+
 	container, err := client.InspectContainer(handle.(*DockerHandle).ContainerID())
 	if err != nil {
 		t.Fatalf("err: %v", err)
@@ -551,12 +536,11 @@ func TestDockerHostNet(t *testing.T) {
 
 	actual := container.HostConfig.NetworkMode
 	if actual != expected {
-		t.Errorf("DNS Network mode doesn't match.\nExpected:\n%s\nGot:\n%s\n", expected, actual)
+		t.Fatalf("Got network mode %q; want %q", expected, actual)
 	}
 }
 
-func TestDockerLabels(t *testing.T) {
-	t.Parallel()
+func TestDockerDriver_Labels(t *testing.T) {
 	task, _, _ := dockerTask()
 	task.Config["labels"] = []map[string]string{
 		map[string]string{
@@ -567,6 +551,8 @@ func TestDockerLabels(t *testing.T) {
 
 	client, handle, cleanup := dockerSetup(t, task)
 	defer cleanup()
+
+	waitForExist(t, client, handle.(*DockerHandle))
 
 	container, err := client.InspectContainer(handle.(*DockerHandle).ContainerID())
 	if err != nil {
@@ -582,14 +568,15 @@ func TestDockerLabels(t *testing.T) {
 	}
 }
 
-func TestDockerDNS(t *testing.T) {
-	t.Parallel()
+func TestDockerDriver_DNS(t *testing.T) {
 	task, _, _ := dockerTask()
 	task.Config["dns_servers"] = []string{"8.8.8.8", "8.8.4.4"}
 	task.Config["dns_search_domains"] = []string{"example.com", "example.org", "example.net"}
 
 	client, handle, cleanup := dockerSetup(t, task)
 	defer cleanup()
+
+	waitForExist(t, client, handle.(*DockerHandle))
 
 	container, err := client.InspectContainer(handle.(*DockerHandle).ContainerID())
 	if err != nil {
@@ -606,7 +593,6 @@ func TestDockerDNS(t *testing.T) {
 }
 
 func TestDockerWorkDir(t *testing.T) {
-	t.Parallel()
 	task, _, _ := dockerTask()
 	task.Config["work_dir"] = "/some/path"
 
@@ -632,12 +618,13 @@ func inSlice(needle string, haystack []string) bool {
 	return false
 }
 
-func TestDockerPortsNoMap(t *testing.T) {
-	t.Parallel()
+func TestDockerDriver_PortsNoMap(t *testing.T) {
 	task, res, dyn := dockerTask()
 
 	client, handle, cleanup := dockerSetup(t, task)
 	defer cleanup()
+
+	waitForExist(t, client, handle.(*DockerHandle))
 
 	container, err := client.InspectContainer(handle.(*DockerHandle).ContainerID())
 	if err != nil {
@@ -650,8 +637,6 @@ func TestDockerPortsNoMap(t *testing.T) {
 		docker.Port(fmt.Sprintf("%d/udp", res)): struct{}{},
 		docker.Port(fmt.Sprintf("%d/tcp", dyn)): struct{}{},
 		docker.Port(fmt.Sprintf("%d/udp", dyn)): struct{}{},
-		// This one comes from the redis container
-		docker.Port("6379/tcp"): struct{}{},
 	}
 
 	if !reflect.DeepEqual(container.Config.ExposedPorts, expectedExposedPorts) {
@@ -683,8 +668,7 @@ func TestDockerPortsNoMap(t *testing.T) {
 	}
 }
 
-func TestDockerPortsMapping(t *testing.T) {
-	t.Parallel()
+func TestDockerDriver_PortsMapping(t *testing.T) {
 	task, res, dyn := dockerTask()
 	task.Config["port_map"] = []map[string]string{
 		map[string]string{
@@ -695,6 +679,8 @@ func TestDockerPortsMapping(t *testing.T) {
 
 	client, handle, cleanup := dockerSetup(t, task)
 	defer cleanup()
+
+	waitForExist(t, client, handle.(*DockerHandle))
 
 	container, err := client.InspectContainer(handle.(*DockerHandle).ContainerID())
 	if err != nil {
@@ -739,15 +725,14 @@ func TestDockerPortsMapping(t *testing.T) {
 	}
 }
 
-func TestDockerUser(t *testing.T) {
-	t.Parallel()
-
+func TestDockerDriver_User(t *testing.T) {
 	task := &structs.Task{
 		Name: "redis-demo",
 		User: "alice",
 		Config: map[string]interface{}{
-			"image":   "redis",
-			"command": "sleep",
+			"image":   "busybox",
+			"load":    []string{"busybox.tar"},
+			"command": "/bin/sleep",
 			"args":    []string{"10000"},
 		},
 		Resources: &structs.Resources{
@@ -767,6 +752,7 @@ func TestDockerUser(t *testing.T) {
 	driverCtx, execCtx := testDriverContexts(task)
 	driver := NewDockerDriver(driverCtx)
 	defer execCtx.AllocDir.Destroy()
+	copyImage(execCtx, task, "busybox.tar", t)
 
 	// It should fail because the user "alice" does not exist on the given
 	// image.
@@ -776,28 +762,17 @@ func TestDockerUser(t *testing.T) {
 		t.Fatalf("Should've failed")
 	}
 
-	msgs := []string{
-		"System error: Unable to find user alice",
-		"linux spec user: Unable to find user alice",
-	}
-	var found bool
-	for _, msg := range msgs {
-		if strings.Contains(err.Error(), msg) {
-			found = true
-			break
-		}
-	}
-	if !found {
+	if !strings.Contains(err.Error(), "alice") {
 		t.Fatalf("Expected failure string not found, found %q instead", err.Error())
 	}
 }
 
 func TestDockerDriver_CleanupContainer(t *testing.T) {
-	t.Parallel()
 	task := &structs.Task{
 		Name: "redis-demo",
 		Config: map[string]interface{}{
 			"image":   "busybox",
+			"load":    []string{"busybox.tar"},
 			"command": "/bin/echo",
 			"args":    []string{"hello"},
 		},
@@ -837,15 +812,14 @@ func TestDockerDriver_CleanupContainer(t *testing.T) {
 	case <-time.After(time.Duration(tu.TestMultiplier()*5) * time.Second):
 		t.Fatalf("timeout")
 	}
-
 }
 
 func TestDockerDriver_Stats(t *testing.T) {
-	t.Parallel()
 	task := &structs.Task{
 		Name: "sleep",
 		Config: map[string]interface{}{
 			"image":   "busybox",
+			"load":    []string{"busybox.tar"},
 			"command": "/bin/sleep",
 			"args":    []string{"100"},
 		},
@@ -858,6 +832,8 @@ func TestDockerDriver_Stats(t *testing.T) {
 
 	_, handle, cleanup := dockerSetup(t, task)
 	defer cleanup()
+
+	waitForExist(t, client, handle.(*DockerHandle))
 
 	go func() {
 		time.Sleep(3 * time.Second)
@@ -884,4 +860,10 @@ func TestDockerDriver_Stats(t *testing.T) {
 		t.Fatalf("timeout")
 	}
 
+}
+
+func copyImage(execCtx *ExecContext, task *structs.Task, image string, t *testing.T) {
+	taskDir, _ := execCtx.AllocDir.TaskDirs[task.Name]
+	dst := filepath.Join(taskDir, allocdir.TaskLocal, image)
+	copyFile(filepath.Join("./test-resources/docker", image), dst, t)
 }